fix responsive, fix logic open modal, remove credentials graphQL client, new button apply whitelist if pool in mode whitelist enabled
