+++ public/locales/en/translation.json
@@ -132,5 +132,6 @@
   "WHITELIST": "Whitelist",
   "ADD_ADDRESS": "Add address",
   "ADD_ADDRESS_TO_WHITELIST_SUCCESSFULLY": "Adding address to whitelist successfully",
-  "SET_WHITELIST": "Set whitelist"
+  "SET_WHITELIST": "Set whitelist",
+  "JOIN_WHITELIST": "Apply Whitelist"
 }
