+++ src/graphQL/client.ts
@@ -6,6 +6,5 @@ const endpoint = config.GRAPHQL_ENDPOINT || 'http://localhost:3000';
 
 const client = new GraphQLClient(endpoint, {
   mode: 'cors',
-  credentials: 'include',
 });
 export default client;
